--- jsettlers.mapcreator/src/jsettlers/exceptionhandler/localization/labels_en_US.properties
@@ -1,29 +1,0 @@
-#Generated by ResourceBundle Editor (http://essiembre.github.io/eclipse-rbe/)
-###############################################################################
-# Copyright (c) 2015
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"),
-# to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense,
-# and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
-# DEALINGS IN THE SOFTWARE.
-###############################################################################
-#
-# ** Tips to translate **
-# May use http://essiembre.github.io/eclipse-rbe/ as editor, makes it a lot easyer
-# Use a dot (.) to group tags by Category, this is common in property files, and then
-# it's displayed as groups in the Translator plugin mentioned above
-
-action.close = Close Dialog
-action.exit  = Exit application
-
-dialog.copy        = Copy text
-dialog.errordesc   = <html>\n<b>There is an unexpected error occured</b>, probably a programming error.<br>\nPlease create a ticket on Github, describe where the error happend,<br>\nand copy the text below in the ticket.<br>\n<b>Thanks!</b>\n</html>
-dialog.header      = JSettlers - Error occured
-dialog.open-failed = Could not open link, please go to %s
-dialog.open-github = Open Github
