+++ jsettlers.mapcreator/src/jsettlers/mapcreator/mapvalidator/tasks/ValidatePlayerStartPosition.java
@@ -22,7 +22,7 @@ public class ValidatePlayerStartPosition extends AbstractValidationTask {
 		for (int player = 0; player < data.getPlayerCount(); player++) {
 			ShortPoint2D point = data.getStartPoint(player);
 			if (players[point.x][point.y] != player) {
-				addErrorMessage("playerstart.text", point, player);
+				addErrorMessage("playerstart.header", point, player);
 			}
 			// mark
 			borders[point.x][point.y] = true;
