+++ jsettlers.tests/.classpath
@@ -14,6 +14,5 @@
 	<classpathentry kind="lib" path="libs/hamcrest-core-1.3.jar"/>
 	<classpathentry kind="lib" path="libs/junit-4.11.jar"/>
 	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER"/>
-	<classpathentry combineaccessrules="false" kind="src" path="/jsettlers.mapcreator"/>
 	<classpathentry kind="output" path="bin"/>
 </classpath>
